@@ -3,9 +3,6 @@ using Application.Interfaces.Repositories;
 using Application.Wrappers;
 using Domain.Entities;
 using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -14,13 +11,16 @@ namespace Application.Features.Contacts.Queries.GetContactById
     public class GetContactByIdQuery : IRequest<Response<Contact>>
     {
         public int Id { get; set; }
+
         public class GetContactByIdQueryHandler : IRequestHandler<GetContactByIdQuery, Response<Contact>>
         {
             private readonly IContactRepositoryAsync _contactRepository;
+
             public GetContactByIdQueryHandler(IContactRepositoryAsync contactRepository)
             {
                 _contactRepository = contactRepository;
             }
+
             public async Task<Response<Contact>> Handle(GetContactByIdQuery query, CancellationToken cancellationToken)
             {
                 var contact = await _contactRepository.GetByIdAsync(query.Id);
